requirements.txt
# V1.0.0 requirements
# --extra-index-url https://download.pytorch.org/whl/cpu
# torch==1.13.1
# torchvision==0.14.1
# requests==2.28.1
# flask==2.1.2
# flask-restx==0.5.1
# flask-cors==3.0.10
# flask-login==0.6.2
# werkzeug==2.1.2
# joblib==1.2.0
# boto3==1.26.8
# python-dotenv==0.21.0
# gunicorn==20.1.0
# scikit-learn==1.2.2

# V1.1.0 and above requirements
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.0.1
torchvision==0.15.2
requests==2.31.0
flask==2.3.3
flask-restx==1.2.0
flask-cors==4.0.0
flask-login==0.6.3
werkzeug==2.3.8
joblib==1.3.2
boto3==1.33.11
python-dotenv==1.0.0
gunicorn==21.2.0
scikit-learn==1.3.0
transformers==4.31.0
sagemaker==2.203.0